Renamed to bearer_token_policy, changed scopes type to Vec<String>, other feedback nits
