+++ sdk/core/azure_core/src/policies/bearer_token_policy.rs
@@ -1,0 +1,50 @@
+use crate::{
+    auth::TokenCredential,
+    headers::AUTHORIZATION,
+    policies::{Policy, PolicyResult},
+    Context, Request,
+};
+use async_trait::async_trait;
+use std::sync::Arc;
+
+#[derive(Debug, Clone)]
+pub struct BearerTokenCredentialPolicy {
+    credential: Arc<dyn TokenCredential>,
+    scopes: Vec<String>,
+}
+
+impl BearerTokenCredentialPolicy {
+    pub fn new<A, B>(credential: Arc<dyn TokenCredential>, scopes: A) -> Self
+    where
+        A: IntoIterator<Item = B>,
+        B: Into<String>,
+    {
+        Self {
+            credential,
+            scopes: scopes.into_iter().map(|s| s.into()).collect(),
+        }
+    }
+}
+
+#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
+#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
+impl Policy for BearerTokenCredentialPolicy {
+    async fn send(
+        &self,
+        ctx: &Context,
+        request: &mut Request,
+        next: &[Arc<dyn Policy>],
+    ) -> PolicyResult {
+        let scopes = self
+            .scopes
+            .iter()
+            .map(|s| s.as_str())
+            .collect::<Vec<&str>>();
+        let access_token = self.credential.get_token(&scopes).await?;
+        let token = access_token.token.secret();
+
+        request.insert_header(AUTHORIZATION, format!("Bearer {token}"));
+
+        next[0].send(ctx, request, &next[1..]).await
+    }
+}
